Request method

+++ web/src/main/java/com/zhefan/yummy/controller/DishesClassController.java
@@ -1,12 +1,15 @@
 package com.zhefan.yummy.controller;
 
-
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.validation.Valid;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -18,6 +21,7 @@ import com.zhefan.yummy.base.BaseController;
 import com.zhefan.yummy.dto.ResponseDTO;
 import com.zhefan.yummy.entity.DishesClass;
 import com.zhefan.yummy.entity.Gerent;
+import com.zhefan.yummy.enums.ResponseEnums;
 import com.zhefan.yummy.param.RequestDishesClass;
 import com.zhefan.yummy.service.DishesClassService;
 import com.zhefan.yummy.util.SessionUtil;
@@ -38,46 +42,53 @@ import io.swagger.annotations.ApiParam;
 @RestController
 @RequestMapping("/dishesClass")
 public class DishesClassController extends BaseController {
-	
+
 	@Autowired
 	private DishesClassService dishesClassService;
-	
+
 	@ApiOperation(value = "列表", notes = "列表")
-	@PostMapping("list")
-	public ResponseDTO<List<DishesClass>> list(@ApiParam("商店ID") Integer shopId, 
-			@ApiParam("菜品类型ID") Integer classId, @ApiParam("类型名") Integer className) {
+	@GetMapping("list")
+	public ResponseDTO<List<DishesClass>> list(@ApiParam("商店ID") Integer shopId, @ApiParam("菜品类型ID") Integer classId,
+			@ApiParam("类型名") Integer className) {
+		if (shopId == null) return ResponseDTO.error(ResponseEnums.SHOP_ID_NULL);
 		Wrapper<DishesClass> wrapper = new EntityWrapper<>();
 		wrapper.eq("shop_id", shopId);
-		if(classId != null) wrapper.eq("id", classId);
-		if(className != null) wrapper.eq("dishes_class_name", className);
+		if (classId != null)
+			wrapper.eq("id", classId);
+		if (className != null)
+			wrapper.eq("dishes_class_name", className);
 		return ResponseDTO.success(dishesClassService.selectList(wrapper));
 	}
-	
+
 	@SuppressWarnings("rawtypes")
 	@ApiOperation(value = "保存", notes = "保存")
 	@PostMapping("save")
-	public ResponseDTO save(@RequestBody RequestDishesClass clas, HttpServletRequest request) {
+	public ResponseDTO save(@Valid @RequestBody RequestDishesClass clas, BindingResult result,
+			HttpServletRequest request) {
 		Gerent gerent = SessionUtil.getLoginBean(request);
 		DishesClass entity = new DishesClass();
 		BeanUtils.copyProperties(clas, entity);
-		if(clas.getId() == null) {
+		if (clas.getId() == null) {
 			entity.setCreatorId(gerent.getId());
 			entity.setCreationTime(getCurrentTime());
 			entity.setCreator(gerent.getNick());
 		}
 		boolean b = dishesClassService.insertOrUpdate(entity);
-		if(!b) return ResponseDTO.error();
+		if (!b)
+			return ResponseDTO.error(ResponseEnums.SAVE_ERROR);
 		return ResponseDTO.success();
 	}
-	
+
 	@SuppressWarnings("rawtypes")
 	@ApiOperation(value = "删除", notes = "删除")
-	@PostMapping("del")
+	@DeleteMapping("del")
 	public ResponseDTO del(@RequestBody List<Integer> ids) {
+		if (ids == null || ids.size() == 0)
+			return ResponseDTO.error(ResponseEnums.DELETE_ERROR);
 		boolean b = dishesClassService.deleteBatchIds(ids);
-		if(!b) return ResponseDTO.error();
+		if (!b)
+			return ResponseDTO.error(ResponseEnums.DELETE_ERROR);
 		return ResponseDTO.success();
 	}
-	
-	
+
 }
