+++ web/src/main/java/com/zhefan/yummy/controller/GerentController.java
@@ -1,15 +1,17 @@
 package com.zhefan.yummy.controller;
 
-
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.validation.Valid;
 
 import org.apache.commons.codec.digest.DigestUtils;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -46,25 +48,32 @@ import lombok.extern.slf4j.Slf4j;
 @RestController
 @RequestMapping("/gerent")
 public class GerentController extends BaseController {
-	
+
 	@Autowired
 	private GerentService gerentService;
-	
+
 	@Autowired
 	private RestTemplate restTemplate;
-	
+
 	@Value("${file.url.del}")
 	private String delUrl;
 	
+	@Value("${file.url.change}")
+	private String changeUrl;
+
 	@ApiOperation(value = "登录", notes = "登录")
 	@SuppressWarnings("rawtypes")
 	@PostMapping("login")
-	public ResponseDTO login(@ApiParam("账号") @RequestParam String name, 
-			@ApiParam("密码") @RequestParam String password, HttpServletRequest request) {
+	public ResponseDTO login(@ApiParam("账号") @RequestParam String name, @ApiParam("密码") @RequestParam String password,
+			HttpServletRequest request) {
+		if (name == null)
+			ResponseDTO.error(ResponseEnums.NAME_ID_NULL);
+		if (password == null)
+			ResponseDTO.error(ResponseEnums.PWD_ID_NULL);
 		Wrapper<Gerent> wrapper = new EntityWrapper<>();
 		wrapper.eq("name", name).eq("password", DigestUtils.md5Hex(password).toUpperCase());
 		Gerent gerent = gerentService.selectOne(wrapper);
-		if(gerent != null) {
+		if (gerent != null) {
 			log.debug(gerent.toString());
 			SessionUtil.setLoginInfo(request, gerent);
 			LinkedMultiValueMap<Object, Object> param = new LinkedMultiValueMap<>();
@@ -74,36 +83,48 @@ public class GerentController extends BaseController {
 		}
 		return ResponseDTO.error(ResponseEnums.LOGIN_ERROR);
 	}
-	
+
 	@SuppressWarnings("rawtypes")
 	@ApiOperation(value = "保存", notes = "保存")
 	@PostMapping("save")
-	public ResponseDTO save(@RequestBody RequestGerent requestGerent, HttpServletRequest request) {
+	public ResponseDTO save(@Valid @RequestBody RequestGerent requestGerent, BindingResult result,
+			HttpServletRequest request) {
 		Gerent gerent = SessionUtil.getLoginBean(request);
 		Gerent entity = new Gerent();
 		BeanUtils.copyProperties(requestGerent, entity);
-		if(requestGerent.getId() == null) {
+		String aImg = entity.getAvatar().replace("temp/", "");
+
+		LinkedMultiValueMap<Object, Object> param = new LinkedMultiValueMap<>();
+		param.add("startFilePath", entity.getAvatar());
+		param.add("endFilePath", aImg);
+		restTemplate.postForObject(changeUrl, param, JSONObject.class);
+		
+		if (requestGerent.getId() == null) {
 			entity.setCreatorId(gerent.getId());
 			entity.setCreationTime(getCurrentTime());
 			entity.setCreator(gerent.getNick());
 		}
 		boolean b = gerentService.insertOrUpdate(entity);
-		if(!b) return ResponseDTO.error();
+		if (!b)
+			return ResponseDTO.error(ResponseEnums.SAVE_ERROR);
 		return ResponseDTO.success();
 	}
-	
+
 	@SuppressWarnings("rawtypes")
 	@ApiOperation(value = "删除", notes = "删除")
-	@PostMapping("del")
+	@DeleteMapping("del")
 	public ResponseDTO del(@RequestBody List<Integer> ids, HttpServletRequest request) {
+		if (ids == null || ids.size() == 0)
+			return ResponseDTO.error(ResponseEnums.DELETE_ERROR);
 		Gerent gerent = SessionUtil.getLoginBean(request);
 		Gerent gerent2 = gerentService.selectById(ids.get(0));
-		if(!"admin".equals(gerent.getName()) && gerent2.getCreatorId().equals(gerent.getId())) {
+		if (!"admin".equals(gerent.getName()) && gerent2.getCreatorId().equals(gerent.getId())) {
 			return ResponseDTO.error("只有管理员和创建本人可操作");
 		}
 		boolean b = gerentService.deleteBatchIds(ids);
-		if(!b) return ResponseDTO.error();
+		if (!b)
+			return ResponseDTO.error(ResponseEnums.DELETE_ERROR);
 		return ResponseDTO.success();
 	}
-	
+
 }
